WIP: UserDetailsManagerImpl, diseñando QUIEN tendrá acceso a los métodos

+++ src/main/java/init/service/UserDetailsManagerImpl.java
@@ -2,10 +2,12 @@ package init.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.provisioning.UserDetailsManager;
 import org.springframework.stereotype.Service;
 
@@ -22,6 +24,9 @@ public class UserDetailsManagerImpl implements UserDetailsManager {
 	
 	@Autowired
 	Mapeador mapeador;
+	
+	@Autowired
+	PasswordEncoder passwordEncoder;
 
 	@Override
 	public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
@@ -56,12 +61,15 @@ public class UserDetailsManagerImpl implements UserDetailsManager {
 
 	@Override
 	public void deleteUser(String username) {
+		if(!usuariosDao.existsByUsername(username)) {
+			throw new UsernameNotFoundException("Usuario no encontrado" + username);
+		}
 		usuariosDao.deleteByUsername(username);
 	}
 
 	@Override
 	public void changePassword(String oldPassword, String newPassword) {
-		//Obtenemos el objeto usuario del Security Context
+		//Obtenemos el objeto Usuario del Security Context
 	    Authentication usuarioActual = SecurityContextHolder.getContext().getAuthentication();
 	    if (usuarioActual == null) {
 	        throw new AccessDeniedException("No hay usuario autenticado");
@@ -70,10 +78,12 @@ public class UserDetailsManagerImpl implements UserDetailsManager {
 	    Usuario usuario = usuariosDao.findByUsername(username);
 	    
 	    //Comprobamos que la contraseña pasada como parámetro sea la misma que figura en la base de datos
-	    if(!oldPassword.equals(usuario.getPassword())) {
-	    	
+	    if(!passwordEncoder.matches(oldPassword, usuario.getPassword())) {
+	        throw new BadCredentialsException("La contraseña proporcionada no es correcta");
 	    }
-		//usuariosDao.updatePassword(username, newPassword);
+	    
+	    //Guardamos la nueva contraseña en la base de datos, asegurándonos de que esté codificada
+	    usuariosDao.updatePassword(username, passwordEncoder.encode(newPassword));
 	}
 
 	@Override
